Move alerting scripts to own folder
