+++ tools/alerts/upload_alerts_to_aws.py
