+++ tools/alerts/validate_alerts.py
